Move IntRange to probability
