+++ src/engine/loot/ItemEntry.ts
@@ -1,6 +1,6 @@
 import {LootEntry} from "@/engine/loot/LootEntry";
 import {ItemId} from "@/game/items/ItemId";
-import {IntRange} from "@/engine/loot/IntRange";
+import {IntRange} from "@/engine/probability/IntRange";
 import {Requirement} from "@/engine/requirements/Requirement";
 import {ItemAmount} from "@/game/items/ItemAmount";
